+++ Aula-12/exercicio1_2/src/br/inatel/cdg/exercicio2/brownie/BrownieNutella.java
@@ -2,7 +2,7 @@ package br.inatel.cdg.exercicio2.brownie;
 
 public class BrownieNutella extends Brownie {
 	
-	public BrownieNutella(String nome, double preco, String sabor) {
+	public BrownieNutella(String nome, int preco, String sabor) {
 		super(nome,preco,sabor);
 	}
